+++ library/entity/networking/build.gradle
@@ -1,0 +1,22 @@
+plugins {
+	id("qsl.module")
+}
+
+qslModule {
+	name = "Quilt Entity Networking API"
+	moduleName = "entity_networking"
+	id = "quilt_entity_networking"
+	description = "An API extending Minecraft's entity data tracking/syncing functionality."
+	library = "entity"
+	moduleDependencies {
+		core {
+			api("qsl_base")
+			api("registry")
+		}
+	}
+	entrypoints {
+		init {
+			values = ["org.quiltmc.qsl.entity.networking.impl.QuiltEntityNetworkingInitializer"]
+		}
+	}
+}
