+++ library/entity/networking/src/main/resources/assets/quilt_entity_networking/icon.png
