+++ library/entity/networking/src/main/resources/quilt_entity_networking.mixins.json
@@ -1,0 +1,14 @@
+{
+  "required": true,
+  "package": "org.quiltmc.qsl.entity.networking.mixin",
+  "compatibilityLevel": "JAVA_17",
+  "mixins": [
+    "TrackedDataHandlerRegistryMixin"
+  ],
+  "client": [
+
+  ],
+  "injectors": {
+    "defaultRequire": 1
+  }
+}
