+++ library/entity/networking/src/testmod/resources/quilt.mod.json
@@ -1,0 +1,24 @@
+{
+  "schema_version": 1,
+  "quilt_loader": {
+    "group": "org.quiltmc.qsl.entity.networking",
+    "id": "qsl_entity_networking_testmod",
+    "version": "1.0.0",
+    "metadata": {
+      "name": "Quilt Tracked Data Handler API Test Mod",
+      "license": "Apache-2.0"
+    },
+    "intermediate_mappings": "net.fabricmc:intermediary",
+    "load_type": "always",
+    "depends": [
+      "quilt_loader",
+      "quilt_entity_networking"
+    ],
+    "entrypoints": {
+      "init": [
+        "org.quiltmc.qsl.entity.networking.test.TrackedDataTestInitializer"
+      ]
+    }
+  },
+  "mixin": "quilt_entity_networking_testmod.mixins.json"
+}
