+++ library/entity/networking/src/testmod/resources/quilt_entity_networking_testmod.mixins.json
@@ -1,0 +1,11 @@
+{
+  "required": true,
+  "package": "org.quiltmc.qsl.entity.networking.test.mixin",
+  "compatibilityLevel": "JAVA_17",
+  "mixins": [
+    "CreeperEntityMixin"
+  ],
+  "injectors": {
+    "defaultRequire": 1
+  }
+}
