--- library/entity/tracked_data/build.gradle
@@ -1,22 +1,0 @@
-plugins {
-	id("qsl.module")
-}
-
-qslModule {
-	name = "Quilt Tracked Data API"
-	moduleName = "tracked_data"
-	id = "quilt_tracked_data"
-	description = "An API extending Minecraft's entity data tracking/syncing functionality."
-	library = "entity"
-	moduleDependencies {
-		core {
-			api("qsl_base")
-			api("registry")
-		}
-	}
-	entrypoints {
-		init {
-			values = ["org.quiltmc.qsl.tracked_data.impl.QuiltTrackedDataInitializer"]
-		}
-	}
-}
