--- library/entity/tracked_data/src/main/resources/quilt_tracked_data.mixins.json
@@ -1,14 +1,0 @@
-{
-  "required": true,
-  "package": "org.quiltmc.qsl.tracked_data.mixin",
-  "compatibilityLevel": "JAVA_17",
-  "mixins": [
-    "TrackedDataHandlerRegistryMixin"
-  ],
-  "client": [
-
-  ],
-  "injectors": {
-    "defaultRequire": 1
-  }
-}
