--- library/entity/tracked_data/src/testmod/resources/quilt_tracked_data_testmod.mixins.json
@@ -1,11 +1,0 @@
-{
-  "required": true,
-  "package": "org.quiltmc.qsl.tracked_data.test.mixin",
-  "compatibilityLevel": "JAVA_17",
-  "mixins": [
-    "CreeperEntityMixin"
-  ],
-  "injectors": {
-    "defaultRequire": 1
-  }
-}
